pyproject.toml
[tool.poetry]
name = "artemis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
keyboard = "^0.13.5"
mss = "^6.1.0"
numpy = "^1.21.4"
opencv-python = "^4.5.4"
pytesseract = "^0.3.8"
PyAutoIt = "^0.6.5"
pynput = "^1.7.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
